+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Select default model
 - Use 'new' after submitting a prompt to immediately start entering a new prompt while the current request
  runs in the background
+- Options for giving files summaries on creation
 
 ### Changed